Move PlayerLayerSet to the mitm module. Should fix #96 for real.
