+++ pyrdp/player/__init__.py
@@ -11,7 +11,6 @@ from pyrdp.player.LiveThread import LiveThread
 from pyrdp.player.LiveWindow import LiveWindow
 from pyrdp.player.MainWindow import MainWindow
 from pyrdp.player.PlayerEventHandler import PlayerEventHandler
-from pyrdp.player.PlayerLayerSet import AsyncIOPlayerLayerSet, TwistedPlayerLayerSet
 from pyrdp.player.QTimerSequencer import QTimerSequencer
 from pyrdp.player.Replay import Replay
 from pyrdp.player.ReplayBar import ReplayBar
